+++ src/templates/object_template.h
@@ -20,6 +20,7 @@
 
 #include <string>
 #include <vector>
+#include <algorithm>
 #include <cstring>
 #include <iostream>
 #include <fstream>